Length-dependent local alignment EVD calculations now working.
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@1388 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/plan7.h
@@ -223,7 +223,7 @@ struct plan7_s {
   float  lambda;		/* EVD lambda                           +*/
   float  kappa;			/* mean length of optimal alignments    +*/
   float  sigma;			/* std dev. of length of opt alignments +*/
-  float  Lbase;			/* seq length that EVD was cali */
+  int    Lbase;			/* seq length that EVD was cali */
 
   int flags;                    /* bit flags indicating state of HMM, valid data +*/
 };
